driving statics menu improvements
