+++ main/utils.cpp
@@ -322,34 +322,18 @@ float wattToMotorCurrent(float watt)
     return wattToAmpere(watt) / 4;
 }
 
-std::string get_current_uptime_string()
-{
-    const auto uptime_time_point = espchrono::utc_clock::now();
-    const auto dateTimeUptime = espchrono::toDateTime(uptime_time_point);
-    std::string out = fmt::format("Up: {:02d}:{:02d}:{:02d}", dateTimeUptime.hour, dateTimeUptime.minute, dateTimeUptime.second);
-    return out;
-}
-
-uint8_t time_to_percent(std::chrono::duration<long, std::ratio<1,1000>> repeat, std::chrono::duration<long, std::ratio<1,1000>> riseTime, std::chrono::duration<long, std::ratio<1,1000>> fullTime, size_t numLeds, bool invert)
+uint8_t time_to_percent(espchrono::milliseconds32 repeat, espchrono::milliseconds32 riseTime, espchrono::milliseconds32 fullTime, size_t numLeds, bool invert)
 {
     const auto now = espchrono::millis_clock::now().time_since_epoch() % repeat;
-    int activated = invert ? numLeds : 0;
     if (now <= riseTime)
     {
         if (invert)
-        {
-            activated = numLeds - ((now*numLeds) / riseTime);
-        }
+            return numLeds - ((now*numLeds) / riseTime);
         else
-        {
-            activated = (now*numLeds) / riseTime;
-        }
+            return (now*numLeds) / riseTime;
     }
     else if (now < riseTime + fullTime)
-    {
-        activated = invert ? 0 : numLeds;
-    }
+        return invert ? 0 : numLeds;
     else
-        activated = invert ? numLeds : 0;
-    return activated;
+        return invert ? numLeds : 0;
 }
